api_mercearia/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
idna==3.4
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
